+++ SoftwareArchitecture/TestYourCalculation/src/main/java/com/akagiyui/testyourcalculation/Range.java
@@ -8,15 +8,15 @@ import lombok.Getter;
 @SuppressWarnings("ClassCanBeRecord")
 @Getter // 生成 getter 方法
 public class Range {
-    private final short start; // 最小值
-    private final short end; // 最大值
+    private final int start; // 最小值
+    private final int end; // 最大值
 
     /**
      * 构造方法
      * @param start 最小值
      * @param end 最大值
      */
-    public Range(short start, short end) {
+    public Range(int start, int end) {
         if (start > end) {
             // 最小值不能大于最大值
             throw new IllegalArgumentException("start > end");
@@ -30,7 +30,7 @@ public class Range {
      * @param value 值
      * @return 是否在范围内
      */
-    public boolean contains(short value) {
+    public boolean contains(int value) {
         return value >= start && value <= end;
     }
 }
